#111 refactor: Extract CellRender to separate components for reuse
