+++ src/components/grid/StatusCellRenderer.module.scss
@@ -1,0 +1,17 @@
+.isActive {
+    background-color: limegreen;
+    color: white;
+    display: inline;
+    padding-inline: 10px;
+    border-radius: 10px;
+    text-align: center;
+}
+
+.isInactive {
+    background-color: indianred;
+    color: white;
+    display: inline;
+    padding-inline: 10px;
+    border-radius: 10px;
+    text-align: center;
+}
