+++ src/features/course-admin/module/ModuleList.module.scss
@@ -5,23 +5,6 @@
     gap: 12px;
     padding-top: 20px;
 }
-.isActive {
-    background-color: limegreen;
-    color: white;
-    display: inline;
-    padding-inline: 10px;
-    border-radius: 10px;
-    text-align: center;
-}
-
-.isInactive {
-    background-color: indianred;
-    color: white;
-    display: inline;
-    padding-inline: 10px;
-    border-radius: 10px;
-    text-align: center;
-}
 
 .spaceBetweenButtonAndSearch {
     display: flex;
